requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
google-generativeai==0.3.2 
Flask
Flask-SQLAlchemy 
Flask-Migrate 